+++ src/core/parsers/youtube.ts
@@ -2,10 +2,10 @@ import { ParserOptions, ParserResult } from "./types";
 
 export function parseYoutube(_options: ParserOptions): ParserResult {
     return {
-        chunkNamer: (chunk) => {
-            const match = chunk.url.match(/\/(\d+?)\/goap/);
+        itemNamer: (item) => {
+            const match = item.url.match(/\/(\d+?)\/goap/);
             if (!match) {
-                throw new Error(`Bad chunk url: ${chunk.url}`);
+                throw new Error(`Bad item url: ${item.url}`);
             }
             return match[1];
         },